requirements.txt
asgiref==3.3.1
Django==3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
mysqlclient==2.0.3
Pygments==2.7.4
pytz==2020.5
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
